+++ gamemodes/terrortown/gamemode/server/sv_main.lua
@@ -1,6 +1,7 @@
 ---
 -- Trouble in Terrorist Town 2
 
+include("ttt2/libraries/spawn.lua")
 include("ttt2/libraries/entity_outputs.lua")
 
 ttt_include("sh_init")
@@ -176,7 +177,7 @@ CHANGED_EQUIPMENT = {}
 -- Called after the gamemode loads and starts.
 -- @hook
 -- @realm server
--- @ref https://wiki.garrysmod.com/page/GM/Initialize
+-- @ref https://wiki.facepunch.com/gmod/GM:Initialize
 -- @local
 function GM:Initialize()
 	MsgN("Trouble In Terrorist Town 2 gamemode initializing...")
@@ -269,7 +270,7 @@ end
 -- @return string The text to be shown in the server browser as the gamemode
 -- @hook
 -- @realm server
--- @ref https://wiki.garrysmod.com/page/GM/GetGameDescription
+-- @ref https://wiki.facepunch.com/gmod/GM:GetGameDescription
 -- @local
 function GM:GetGameDescription()
 	return self.Name
@@ -284,7 +285,7 @@ end
 -- the client will receive it as NULL entity.
 -- @hook
 -- @realm server
--- @ref https://wiki.garrysmod.com/page/GM/InitPostEntity
+-- @ref https://wiki.facepunch.com/gmod/GM:InitPostEntity
 -- @local
 function GM:InitPostEntity()
 	self:InitCvars()
@@ -390,7 +391,7 @@ end
 -- Called after the gamemode has loaded
 -- @hook
 -- @realm server
--- @ref https://wiki.garrysmod.com/page/GM/PostGamemodeLoaded
+-- @ref https://wiki.facepunch.com/gmod/GM:PostGamemodeLoaded
 -- @local
 function GM:PostGamemodeLoaded()
 
@@ -738,7 +739,7 @@ end
 -- Called right before the map cleans up (usually because @{game.CleanUpMap} was called)
 -- @hook
 -- @realm server
--- @ref https://wiki.garrysmod.com/page/GM/PreCleanupMap
+-- @ref https://wiki.facepunch.com/gmod/GM:PreCleanupMap
 -- @local
 function GM:PreCleanupMap()
 	ents.TTT.FixParentedPreCleanup()
@@ -750,7 +751,7 @@ end
 -- Called right after the map has cleaned up (usually because game.CleanUpMap was called)
 -- @hook
 -- @realm server
--- @ref https://wiki.garrysmod.com/page/GM/PostCleanupMap
+-- @ref https://wiki.facepunch.com/gmod/GM:PostCleanupMap
 -- @local
 function GM:PostCleanupMap()
 	ents.TTT.FixParentedPostCleanup()
@@ -963,7 +964,11 @@ function PrepareRound()
 	local plys = player.GetAll()
 
 	for i = 1, #plys do
-		plys[i]:SetTargetPlayer(nil)
+		local ply = plys[i]
+
+		ply:SetTargetPlayer(nil)
+		ply:ResetRoundDeathCounter()
+		ply:SetActiveInRound(false)
 	end
 
 	-- Tell hooks and map we started prep
@@ -1043,7 +1048,7 @@ function SpawnWillingPlayers(dead_only)
 		end
 	else
 		-- wave method
-		local num_spawns = #GetSpawnEnts()
+		local num_spawns = #spawn.GetPlayerSpawnEntities()
 		local to_spawn = {}
 
 		for _, ply in RandomPairs(plys) do
@@ -1176,6 +1181,17 @@ function BeginRound()
 	GAMEMODE:UpdatePlayerLoadouts() -- needs to happen when round_active
 
 	ARMOR:InitPlayerArmor()
+
+	local plys = player.GetAll()
+
+	for i = 1, #plys do
+		local ply = plys[i]
+
+		ply:ResetRoundDeathCounter()
+
+		-- a player should be considered "was active in round" if they received a role
+		ply:SetActiveInRound(ply:Alive() and ply:IsTerror())
+	end
 
 	hook.Call("TTTBeginRound", GAMEMODE)
 
@@ -1337,7 +1353,7 @@ function GM:TTTCheckForWin()
 		local v = plys[i]
 		local tm = v:GetTeam()
 
-		if (v:IsTerror() or v.forceRevive) and not v:GetSubRoleData().preventWin and tm ~= TEAM_NONE then
+		if (v:IsTerror() or v:IsBlockingRevival()) and not v:GetSubRoleData().preventWin and tm ~= TEAM_NONE then
 			alive[#alive + 1] = tm
 		end
 	end