+++ gamemodes/terrortown/gamemode/shared/lang/italian.lua
@@ -1290,5 +1290,12 @@ L.hud_health = "Vita"
 --L.item_no_drown_damage = "No Drowning Damage"
 --L.item_no_drown_damage_desc = [[Makes you immune to drowning damage.]]
 
+-- 2020-04-30
+--L.message_revival_canceled = "Revival canceled."
+--L.message_revival_failed = "Revival failed."
+--L.message_revival_failed_missing_body = "You have not been revived because your corpse no longer exists."
+--L.hud_revival_title = "Time left until revival:"
+--L.hud_revival_time = "{time}s"
+
 -- 2020-05-03
 --L.door_destructible = "Door is destructible ({health}HP)"